#124 [feat] : SheetItem PC ver Home에서 사용하기 위해 분리
